requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
docker==6.1.3
fonttools==4.42.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.33
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
more-itertools==10.1.0
multidict==6.0.4
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
pypeln==0.4.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
smmap==5.0.0
stopit==1.1.2
urllib3==2.0.4
websocket-client==1.6.2
yarl==1.9.2
